fixing positioning of searchbar and icon
