+++ src/components/Search/SearchSection.js
@@ -1,17 +1,37 @@
 import React from 'react';
-import { View, Text } from 'react-native';
+import { View, Text, Button } from 'react-native';
 import SearchBar from './SearchBar';
-
+import Icon from'react-native-vector-icons/FontAwesome';
+const glass = (<Icon name ="search" size={20} color="#FA8D62"/>)
 
 const SearchSection = (props) => {
 
     return(
-        <View>
+        <View style={{flexDirection: 'row'}}>
+        <View style={TextInputStyle}>
             {props.children}
-        <SearchBar/>
+        <SearchBar />
         </View>
+        <View>
+
+            </View>
+            </View>
     )
 
+}
+const TextInputStyle = {
+    
+    height: 40,
+    // borderColor: 'black',
+    // borderWidth: 1,
+    width: "70%",
+    marginLeft: 15,
+    // marginRight: 80,
+    backgroundColor: 'white',
+    borderRadius: 5,
+    paddingLeft: 15,
+    marginBottom: 20,
+    marginTop: 10
 }
